#39 move data store to common
